Move spring related components to sub directory
