--- backend/src/main/java/com/sprawler/cache/config/CacheConfig.java
@@ -1,17 +1,0 @@
-package com.sprawler.cache.config;
-
-import org.springframework.cache.CacheManager;
-import org.springframework.cache.annotation.EnableCaching;
-import org.springframework.cache.concurrent.ConcurrentMapCacheManager;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-
-@Configuration
-@EnableCaching
-public class CacheConfig {
-
-    @Bean("cacheManager")
-    public CacheManager cacheManager() {
-        return new ConcurrentMapCacheManager("cache");
-    }
-}
