--- backend/src/main/java/com/sprawler/email/AttachmentEmailRequest.java
@@ -1,8 +1,0 @@
-package com.sprawler.email;
-
-public record AttachmentEmailRequest(String recipient,
-                                     String subject,
-                                     String text,
-                                     String attachmentName,
-                                     String attachment) {
-}
