--- backend/src/main/java/com/sprawler/email/BasicEmailRequest.java
@@ -1,4 +1,0 @@
-package com.sprawler.email;
-
-public record BasicEmailRequest(String recipient, String subject, String text) {
-}
