--- backend/src/main/java/com/sprawler/hibernate/book/Book.java
@@ -1,22 +1,0 @@
-package com.sprawler.hibernate.book;
-
-import jakarta.persistence.Entity;
-import jakarta.persistence.GeneratedValue;
-import jakarta.persistence.Id;
-import lombok.Getter;
-import lombok.Setter;
-import lombok.NoArgsConstructor;
-
-
-@Entity
-@NoArgsConstructor
-@Getter
-@Setter
-public class Book {
-    @Id
-    @GeneratedValue
-    private Long id;
-    private String title;
-    private String author;
-
-}
