--- backend/src/main/java/com/sprawler/hibernate/commons/HibernateService.java
@@ -1,17 +1,0 @@
-package com.sprawler.hibernate.commons;
-
-import java.util.List;
-import java.util.Optional;
-
-public interface HibernateService<T> {
-
-    List<T> getAllEntities();
-
-    Optional<T> getEntityById(Long id);
-
-    T createEntity(T entity);
-
-    T updateEntity(Long id, T updatedEntity);
-
-    void deleteEntity(Long id);
-}
