--- backend/src/main/java/com/sprawler/jms/JmsReceiver.java
@@ -1,14 +1,0 @@
-package com.sprawler.jms;
-
-import org.springframework.jms.annotation.JmsListener;
-import org.springframework.stereotype.Service;
-
-@Service("mqJmsReceiver")
-public class JmsReceiver {
-
-    @JmsListener(destination = "myDestination")
-    public void sampleJmsListenerMethod(String incomingMessage) {
-        System.out.println(incomingMessage);
-    }
-    
-}
