--- backend/src/main/java/com/sprawler/jms/JmsRequest.java
@@ -1,4 +1,0 @@
-package com.sprawler.jms;
-
-public record JmsRequest(String queue, String message) {
-}
