--- backend/src/main/java/com/sprawler/jms/JmsResponse.java
@@ -1,4 +1,0 @@
-package com.sprawler.jms;
-
-public record JmsResponse(String message) {
-}
