--- backend/src/main/java/com/sprawler/redis/RedisRequest.java
@@ -1,4 +1,0 @@
-package com.sprawler.redis;
-
-public record RedisRequest(String key, String value) {
-}
