--- backend/src/main/java/com/sprawler/redis/RedisResponse.java
@@ -1,4 +1,0 @@
-package com.sprawler.redis;
-
-public record RedisResponse(String key, String value) {
-}
