--- backend/src/main/java/com/sprawler/rest/CrudRequest.java
@@ -1,5 +1,0 @@
-package com.sprawler.rest;
-
-public record CrudRequest (String name) {
-
-}
