--- backend/src/main/java/com/sprawler/rest/CrudResponse.java
@@ -1,5 +1,0 @@
-package com.sprawler.rest;
-
-public record CrudResponse(long id, String content) {
-
-}
