--- backend/src/main/java/com/sprawler/rest/HttpMethod.java
@@ -1,9 +1,0 @@
-package com.sprawler.rest;
-
-public enum HttpMethod {
-    GET,
-    POST,
-    PUT,
-    PATCH,
-    DELETE
-}
