+++ backend/src/main/java/com/sprawler/spring/concurrency/config/ConcurrencyConfig.java
@@ -1,0 +1,29 @@
+package com.sprawler.spring.concurrency.config;
+
+import org.springframework.cache.CacheManager;
+import org.springframework.cache.concurrent.ConcurrentMapCacheManager;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.scheduling.annotation.EnableAsync;
+import org.springframework.scheduling.annotation.EnableScheduling;
+import org.springframework.scheduling.concurrent.ThreadPoolTaskExecutor;
+
+import java.util.concurrent.Executor;
+
+@Configuration
+@EnableAsync
+@EnableScheduling
+public class ConcurrencyConfig {
+
+    @Bean("taskExecutor")
+    public Executor taskExecutor() {
+        ThreadPoolTaskExecutor executor = new ThreadPoolTaskExecutor();
+
+        executor.setCorePoolSize(5); // Minimum number of threads
+        executor.setMaxPoolSize(7); // Maximum number of threads
+        executor.setQueueCapacity(100); // Queue size
+        executor.initialize();
+
+        return executor;
+    }
+}
