+++ backend/src/main/java/com/sprawler/spring/email/AttachmentEmailRequest.java
@@ -1,0 +1,8 @@
+package com.sprawler.spring.email;
+
+public record AttachmentEmailRequest(String recipient,
+                                     String subject,
+                                     String text,
+                                     String attachmentName,
+                                     String attachment) {
+}
