+++ backend/src/main/java/com/sprawler/spring/email/BasicEmailRequest.java
@@ -1,0 +1,4 @@
+package com.sprawler.spring.email;
+
+public record BasicEmailRequest(String recipient, String subject, String text) {
+}
