+++ backend/src/main/java/com/sprawler/spring/hibernate/book/BookRepository.java
@@ -1,0 +1,8 @@
+package com.sprawler.spring.hibernate.book;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository("bookRepository")
+public interface BookRepository extends JpaRepository<Book, Long> {
+}
