+++ backend/src/main/java/com/sprawler/spring/hibernate/commons/HibernateService.java
@@ -1,0 +1,17 @@
+package com.sprawler.spring.hibernate.commons;
+
+import java.util.List;
+import java.util.Optional;
+
+public interface HibernateService<T> {
+
+    List<T> getAllEntities();
+
+    Optional<T> getEntityById(Long id);
+
+    T createEntity(T entity);
+
+    T updateEntity(Long id, T updatedEntity);
+
+    void deleteEntity(Long id);
+}
