+++ backend/src/main/java/com/sprawler/spring/jms/JmsRequest.java
@@ -1,0 +1,4 @@
+package com.sprawler.spring.jms;
+
+public record JmsRequest(String queue, String message) {
+}
