+++ backend/src/main/java/com/sprawler/spring/jms/JmsResponse.java
@@ -1,0 +1,4 @@
+package com.sprawler.spring.jms;
+
+public record JmsResponse(String message) {
+}
