+++ backend/src/main/java/com/sprawler/spring/redis/RedisRequest.java
@@ -1,0 +1,4 @@
+package com.sprawler.spring.redis;
+
+public record RedisRequest(String key, String value) {
+}
