+++ backend/src/main/java/com/sprawler/spring/redis/RedisResponse.java
@@ -1,0 +1,4 @@
+package com.sprawler.spring.redis;
+
+public record RedisResponse(String key, String value) {
+}
