+++ backend/src/main/java/com/sprawler/spring/rest/CrudRequest.java
@@ -1,0 +1,5 @@
+package com.sprawler.spring.rest;
+
+public record CrudRequest (String name) {
+
+}
