+++ backend/src/main/java/com/sprawler/spring/rest/CrudResponse.java
@@ -1,0 +1,5 @@
+package com.sprawler.spring.rest;
+
+public record CrudResponse(long id, String content) {
+
+}
