+++ backend/src/main/java/com/sprawler/spring/rest/HttpMethod.java
@@ -1,0 +1,9 @@
+package com.sprawler.spring.rest;
+
+public enum HttpMethod {
+    GET,
+    POST,
+    PUT,
+    PATCH,
+    DELETE
+}
